Rename fcconfigs to pycandy
