+++ server/pycandy/__init__.py
@@ -1,0 +1,3 @@
+"""
+A Python wrapper library for Open Pixel Control.
+"""
