+++ server/pycandy/solid_color.py
